Refactor vendor-aster application structure and services
- Removed obsolete files: cli.ts, order-spot.ts, order.ts, pending-orders.ts, sapi.ts.
- Updated index.ts to import new service modules from the services directory.
- Introduced new services for account actions, order actions, and market data.
- Implemented credential-based actions for submitting and canceling orders.
- Added support for fetching account information for both spot and perpetual accounts.
- Created market services for handling interest rates, products, and quotes.
- Enhanced order handling with separate functions for spot and perpetual orders.
- Updated utility functions and removed unused HMAC function.
- Revised API documentation to reflect new action handler types and services.
