+++ apps/vendor-aster/src/api/public-api.ts
@@ -1,0 +1,68 @@
+const BASE_URL = 'https://fapi.asterdex.com';
+
+const request = async <T>(method: string, endpoint: string, params: any = {}): Promise<T> => {
+  const url = new URL(BASE_URL);
+  url.pathname = endpoint;
+  for (const [key, value] of Object.entries(params)) {
+    if (value === undefined) continue;
+    url.searchParams.set(key, `${value}`);
+  }
+
+  console.info(url.toString());
+  const res = await fetch(url.toString(), {
+    method,
+  }).then((response) => response.json());
+  if (res.code && res.code !== 0) {
+    throw JSON.stringify(res);
+  }
+  return res;
+};
+
+const createApi =
+  <TReq, TRes>(method: string, endpoint: string) =>
+  (params: TReq) =>
+    request<TRes>(method, endpoint, params);
+
+/**
+ * 获取资金费率历史
+ *
+ * https://github.com/asterdex/api-docs/blob/master/aster-finance-futures-api_CN.md#%E6%9F%A5%E8%AF%A2%E8%B5%84%E9%87%91%E8%B4%B9%E7%8E%87%E5%8E%86%E5%8F%B2
+ */
+export const getFApiV1FundingRate = createApi<
+  {
+    symbol?: string;
+    startTime?: number;
+    endTime?: number;
+    limit?: number;
+  },
+  {
+    symbol: string;
+    fundingRate: string;
+    fundingTime: number;
+  }[]
+>('GET', '/fapi/v1/fundingRate');
+
+/**
+ * 获取交易对信息
+ *
+ * https://github.com/asterdex/api-docs/blob/master/aster-finance-futures-api_CN.md#%E4%BA%A4%E6%98%93%E5%AF%B9%E4%BF%A1%E6%81%AF
+ */
+export const getFApiV1ExchangeInfo = createApi<
+  {},
+  {
+    symbols: {
+      symbol: string;
+      status: 'TRADING' | 'BREAK' | 'HALT';
+      baseAsset: string;
+      quoteAsset: string;
+      pricePrecision: number;
+      quantityPrecision: number;
+      baseAssetPrecision: number;
+      quotePrecision: number;
+      filters: {
+        filterType: string;
+        [key: string]: any;
+      }[];
+    }[];
+  }
+>('GET', '/fapi/v1/exchangeInfo');
