--- apps/vendor-aster/src/cli.ts
@@ -1,2 +1,0 @@
-#!/usr/bin/env node
-import './index';
