+++ apps/vendor-aster/src/index.ts
@@ -1,8 +1,6 @@
-import './account';
-import './account-spot';
-import './order';
-import './order-spot';
-import './interest_rate';
-import './product'; // Import the new product service
-import './pending-orders';
-import './quote';
+import './services/markets/interest_rate';
+import './services/markets/product';
+import './services/markets/quote';
+import './services/account-actions-with-credential';
+import './services/legacy';
+import './services/order-actions-with-credential';
