--- apps/vendor-aster/src/pending-orders.ts
@@ -1,71 +1,0 @@
-import { IOrder, providePendingOrdersService } from '@yuants/data-order';
-import { Terminal } from '@yuants/protocol';
-import { encodePath } from '@yuants/utils';
-import { ACCOUNT_ID } from './account';
-import { getFApiV1OpenOrders } from './api';
-
-const terminal = Terminal.fromNodeEnv();
-
-type OrderDirection = 'OPEN_LONG' | 'OPEN_SHORT' | 'CLOSE_LONG' | 'CLOSE_SHORT';
-
-interface IAsterOpenOrder {
-  orderId: number;
-  symbol: string;
-  side: 'BUY' | 'SELL';
-  positionSide?: 'BOTH' | 'LONG' | 'SHORT';
-  reduceOnly?: boolean;
-  closePosition?: boolean;
-  type: string;
-  origQty: string;
-  executedQty: string;
-  price: string;
-  avgPrice: string;
-  status: string;
-  updateTime: number;
-}
-
-const resolveOrderDirection = (asterOrder: IAsterOpenOrder): OrderDirection => {
-  const reduceOnly = asterOrder.reduceOnly || asterOrder.closePosition;
-
-  if (asterOrder.positionSide === 'LONG') {
-    return asterOrder.side === 'BUY' ? 'OPEN_LONG' : 'CLOSE_LONG';
-  }
-  if (asterOrder.positionSide === 'SHORT') {
-    return asterOrder.side === 'BUY' ? 'CLOSE_SHORT' : 'OPEN_SHORT';
-  }
-  if (reduceOnly) {
-    return asterOrder.side === 'BUY' ? 'CLOSE_SHORT' : 'CLOSE_LONG';
-  }
-  return asterOrder.side === 'BUY' ? 'OPEN_LONG' : 'OPEN_SHORT';
-};
-
-providePendingOrdersService(
-  terminal,
-  ACCOUNT_ID,
-  async () => {
-    const orders = (await getFApiV1OpenOrders({})) as IAsterOpenOrder[];
-    return orders.map((order) => {
-      const volume = Number(order.origQty);
-      const tradedVolume = Number(order.executedQty);
-      const price = Number(order.price);
-      const avgPrice = Number(order.avgPrice);
-
-      const mapped: IOrder = {
-        order_id: `${order.orderId}`,
-        account_id: ACCOUNT_ID,
-        product_id: encodePath('PERPETUAL', order.symbol),
-        order_type: order.type,
-        order_direction: resolveOrderDirection(order),
-        volume: Number.isFinite(volume) ? volume : 0,
-        price: Number.isFinite(price) ? price : undefined,
-        submit_at: order.updateTime,
-        traded_volume: Number.isFinite(tradedVolume) ? tradedVolume : undefined,
-        traded_price: Number.isFinite(avgPrice) && avgPrice > 0 ? avgPrice : undefined,
-        order_status: order.status,
-      };
-
-      return mapped;
-    });
-  },
-  { auto_refresh_interval: 2000 },
-);
