+++ apps/vendor-aster/src/services/markets/quote.ts
@@ -1,18 +1,20 @@
 import { createCache } from '@yuants/cache';
+import type { IQuote } from '@yuants/data-quote';
 import { Terminal } from '@yuants/protocol';
 import { writeToSQL } from '@yuants/sql';
 import { decodePath, encodePath } from '@yuants/utils';
 import { defer, filter, from, map, mergeMap, repeat, retry, shareReplay } from 'rxjs';
-import type { IQuote } from '../../../libraries/data-quote/lib';
-import { getFApiV1OpenInterest, getFApiV1TickerPrice } from './api';
+import { getDefaultCredential, getFApiV1OpenInterest, getFApiV1TickerPrice } from '../../api/private-api';
 
 const terminal = Terminal.fromNodeEnv();
 const OPEN_INTEREST_TTL = process.env.OPEN_INTEREST_TTL ? Number(process.env.OPEN_INTEREST_TTL) : 120_000;
 
+const credential = getDefaultCredential();
+
 const openInterestCache = createCache<string>(
   async (symbol: string) => {
     try {
-      const data = await getFApiV1OpenInterest({ symbol });
+      const data = await getFApiV1OpenInterest(credential, { symbol });
       return data.openInterest;
     } catch (error) {
       console.warn('getFApiV1OpenInterest failed', symbol, error);
@@ -22,7 +24,7 @@ const openInterestCache = createCache<string>(
   { expire: OPEN_INTEREST_TTL },
 );
 
-const quote$ = defer(() => getFApiV1TickerPrice({})).pipe(
+const quote$ = defer(() => getFApiV1TickerPrice(credential, {})).pipe(
   mergeMap((tickers) => tickers || []),
   mergeMap(
     (ticker) =>
