+++ common/changes/@yuants/vendor-aster/2025-11-16-22-06.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@yuants/vendor-aster",
+      "comment": "rewrite",
+      "type": "minor"
+    }
+  ],
+  "packageName": "@yuants/vendor-aster"
+}
